ios-wda/requirements.txt
black==22.12.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.13
facebook-wda==1.4.6
idna==3.4
mypy-extensions==0.4.3
numpy==1.24.1
opencv-python==4.7.0.68
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.2
py==1.11.0
requests==2.28.1
retry==0.9.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
wrapt==1.14.1
